use invoke instead of send
