--- src/general/register/tools.ts
@@ -1,9 +1,0 @@
-import { Globals } from '@Globals'
-
-export function getSuccessfulEvent(baseName: string) {
-    return baseName + "_successful"
-}
-
-export function getErrorEvent(baseName: string) {
-    return baseName + "_error"
-}
